compose: move RecipientSelectView into appropriate package
